==> notebooks/01_Bases_MELI_Challenge.ipynb <==
# 📦 Mercado Libre Data Engineer Challenge
### Este notebook genera datos simulados.

Instalación de paquetes

!pip install Faker

Importaciones y configuraciones globales

from faker import Faker
import pandas as pd
import random
import unicodedata
from datetime import datetime, timedelta
#Se obtiene ejemplos de Argentina
fake = Faker('es_AR')

Fuciones útiles

# Generar mail de acuerdo al nombre
def generar_email(nombre, apellido):
    dominio = random.choice(lista_dominios)
    return f"{limpiar(nombre)}.{limpiar(apellido)}{dominio}"

# Normalizar y limpar el mail
def limpiar(texto):
    return unicodedata.normalize('NFKD', texto.split()[0]).encode('ascii', 'ignore').decode('utf-8').lower()

Generar Clientes

# Dominios de mail
lista_dominios = ["@hotmail.com", "@gmail.com", "@outlook.com", "@yahoo.com", "@icloud.com"]

def generar_customers(n):
    customers = []
    for i in range(1, n + 1):
      nombre = fake.first_name()
      apellido = fake.last_name()
      customers.append({
          "customer_id": i,
          "nombre": nombre,
          "apellido": apellido,
          "email": generar_email(nombre, apellido),
          "sexo": random.choice(["M", "F"]),
          "direccion": fake.address().replace("\n", ", "),
          "nacimiento": fake.date_of_birth(minimum_age=18, maximum_age=85),
          "telefono": fake.phone_number()
      })
    return customers

Generar Categorías

# Genera base aleatoria de Categories
def generar_categories():
    categories = [
        {"category_id": 1, "nombre_categoria": "Celulares", "path_item": "Tecnología > Celulares y Teléfonos > Celulares y Smartphones"},
        {"category_id": 2, "nombre_categoria": "Notebooks", "path_item": "Tecnología > Computadoras > Notebooks"},
        {"category_id": 3, "nombre_categoria": "Cafeteras", "path_item": "Hogar > Electrodomésticos y Aires Ac. > Pequeños Electrodomésticos> Para Cocina > Cafeteras"},
    ]
    return categories

Generar Items

# Genera base aleatoria de Items
def generar_items(q=15, lista_clientes=None, por=5):
    if not lista_clientes or not isinstance(lista_clientes, list):
        raise ValueError("Debés pasar una lista válida de clientes.")

    # Lista que guarda los items
    items = []

    # Productos Principales por Categoria
    lista_celulares = ["Samsung Galaxy A15", "Xiaomi Redmi Note 13 Pro", "Motorola Moto G24 Power", "iPhone 16 Pro", "iPhone 17 Pro Max"]
    lista_notebooks = ["MacBook Pro 13", "HP 255 G10", "Lenovo Ideapad Slim 3", "Lenovo Ideapad Slim 3", "Samsung Galaxy Np750 Book3"]
    lista_cafeteras = ["Dolce Gusto Piccolo XS", " Smartchef Pe-ce5002 Automática Expreso P1", "Dolce Gusto Genio S", "Nespresso Essenza Mini D", "Peabody Smartchef Pe-ct4207"]

    # Obtener todos los customer_id desde la lista
    all_customer_ids = [c['customer_id'] for c in lista_clientes]

    # Determinar el %(parámetro por) de los clientes como vendedores
    total_vendedores = max(1, int(len(all_customer_ids) * por //100 ))
    seller_ids = random.sample(all_customer_ids, total_vendedores)

    # Mapear category_id a sus respectivas listas
    productos_por_categoria = {
        1: lista_celulares,
        2: lista_notebooks,
        3: lista_cafeteras
    }

    for i in range(1, q + 1):
        categoria_id = random.choice([1, 2, 3])
        titulo = random.choice(productos_por_categoria[categoria_id])

        items.append({
            "item_id": i,
            "titulo": titulo,
            "precio": random.randint(100, 1000),
            "status": random.choice(["active", "inactive"]),
            "category_id": categoria_id,
            "seller_id": random.choice(seller_ids)
        })

    return items

Generar Orders

# Genera base aleatoria de Orders
def generar_orders(o=4000, lista_items=None, lista_clientes=None):
    if not lista_items or not lista_clientes:
        raise ValueError("Debés pasar listas válidas de ítems y clientes.")

    orders = []
    fecha_base = datetime(2020, 1, 1)

    for i in range(1, o + 1):
        cantidad = random.randint(1, 4)
        order_date = fecha_base + timedelta(days=random.randint(0, 364))

        item = random.choice(lista_items)
        buyer = random.choice(lista_clientes)

        # Evitar que buyer y seller sean la misma persona (opcional)
        while buyer["customer_id"] == item["seller_id"]:
            buyer = random.choice(lista_clientes)

        orders.append({
            "order_id": i,
            "item_id": item["item_id"],
            "seller_id": item["seller_id"],
            "buyer_id": buyer["customer_id"],
            "cantidad": cantidad,
            "precio": item["precio"],
            "order_date": order_date.strftime('%Y-%m-%d')
        })

    return orders

n = int(input("Ingrese la cantidad de usuarios que va a tener la base: "))

customers = generar_customers(n)

items = generar_items(q=1500, lista_clientes=customers,por=5)

categories = generar_categories()

orders = generar_orders(o=13000, lista_clientes=customers, lista_items=items)

Se crean Dataframes a partir de las Listas de entidades para explorar la información generada

df_customers = pd.DataFrame(customers)
df_items = pd.DataFrame(items)
df_categories = pd.DataFrame(categories)
df_orders = pd.DataFrame(orders)

df_customers.sample(10)

df_items

df_categories

df_orders

df_orders.sort_values(by='order_date', ascending=False)

Generar base para tener una tabla de Items Histórico por día

# Generar los últimos 5 días del 2020
start_date = datetime(2020, 12, 27)
end_date = datetime(2020, 12, 31)
delta = end_date - start_date

# Inicializar lista para snapshots
snapshots = []

for day in range(delta.days + 1):  # incluye el 31/12
    snapshot_date = (start_date + timedelta(days=day)).strftime("%Y-%m-%d")

    for _, row in df_items.iterrows():
        snapshots.append({
            "item_id": row["item_id"],
            "snapshot_date": snapshot_date,
            "precio": int(row["precio"] * random.uniform(0.95, 1.05)),
            "status": random.choice(["active", "inactive"])
        })

# Crear el DataFrame y exportarlo
df_hist = pd.DataFrame(snapshots)

df_hist

Guardo las bases en formato csv

pd.DataFrame(customers).to_csv("customers.csv", index=False)
pd.DataFrame(items).to_csv("items.csv", index=False)
pd.DataFrame(categories).to_csv("categories.csv", index=False)
pd.DataFrame(orders).to_csv("orders.csv", index=False)
pd.DataFrame(df_hist).to_csv("item_history.csv", index=False)
