==> notebooks/02_Respuestas_Negocio_MELI_Challenge_.ipynb <==
#  PySpark SQL - Mercado Libre Challenge
Este notebook carga los archivos CSV generados en el primer paso y ejecuta consultas sobre Spark SQL para resolver las consignas del challenge.

##  Instalar PySpark en Google Colab

!pip install pyspark

## Crear SparkSession

import os
from pyspark.sql import SparkSession

spark = SparkSession.builder.appName("MeliChallenge").getOrCreate()

##  Cargar archivos CSV

from google.colab import files

# customers.csv, items.csv, orders.csv y categories.csv
uploaded = files.upload()

##  Leer los archivos CSV como DataFrames de Spark

df_customers = spark.read.option("header", "true").option("inferSchema", "true").csv("customers.csv")
df_items = spark.read.option("header", "true").option("inferSchema", "true").csv("items.csv")
df_orders = spark.read.option("header", "true").option("inferSchema", "true").csv("orders.csv")
df_categories = spark.read.option("header", "true").option("inferSchema", "true").csv("categories.csv")

##  Crear tablas temporales SQL

df_customers.createOrReplaceTempView("Customer")
df_items.createOrReplaceTempView("Item")
df_orders.createOrReplaceTempView("Order")
df_categories.createOrReplaceTempView("Category")

## Usuarios que cumplen años hoy y vendieron más de $1500 en enero 2020

spark.sql("""
SELECT
  c.customer_id, c.nombre, c.apellido, c.nacimiento, SUM(o.precio * o.cantidad) AS total_ventas
FROM Customer c
JOIN Order o ON c.customer_id = o.seller_id
WHERE MONTH(o.order_date) = 1 AND YEAR(o.order_date) = 2020
  AND DATE_FORMAT(c.nacimiento, 'MM-dd') = DATE_FORMAT(CURRENT_DATE(), 'MM-dd')
GROUP BY 1, 2, 3, 4
HAVING total_ventas > 1500
""").show()

Genero consulta de control rápido para verificar la query anterior

spark.sql("""
SELECT
  order_id,	item_id,	seller_id	,	cantidad,	precio,	order_date,
  SUM(precio * cantidad) AS total_ventas
FROM Order
WHERE MONTH(order_date) = 1 AND YEAR(order_date) = 2020
  AND seller_id = 2120
  GROUP BY 1,2,3,4,5,6
""").show()

## Por cada mes del 2020, se solicita el top 5 de usuarios que más vendieron($) en la categoría Celulares. Se requiere el mes y año de análisis, nombre y apellido del vendedor, cantidad de ventas realizadas, cantidad de productos vendidos y el monto total transaccionado.

# Consulta 2 - Top 5 vendedores mensuales en Celulares
spark.sql("""
WITH ventas AS (
  SELECT
    MONTH(o.order_date) AS mes,
    c.customer_id,
    c.nombre,
    c.apellido,
    COUNT(o.order_id) AS cantidad_ventas,
    SUM(o.cantidad) AS productos_vendidos,
    SUM(o.precio * o.cantidad) AS total
  FROM order o
  JOIN item i ON o.item_id = i.item_id
  JOIN category cat ON i.category_id = cat.category_id
  JOIN customer c ON c.customer_id = o.seller_id
  WHERE
    YEAR(o.order_date) = 2020 AND
    LOWER(cat.nombre_categoria) LIKE '%celulares%'
  GROUP BY
    MONTH(o.order_date), c.customer_id, c.nombre, c.apellido
),
ranked AS (
  SELECT *,
         ROW_NUMBER() OVER (PARTITION BY mes ORDER BY total DESC) AS ranking
  FROM ventas
)
SELECT *
FROM ranked
WHERE ranking <= 5
""").show()
