==> notebooks/03_APIS_MELI_Challenge.ipynb <==
# 📦 Mercado Libre Data Engineer Challenge
Este notebook carga datos simulados y ejecuta consultas SQL sobre una base de datos SQLite.

# librerias
import pandas as pd
import json
import requests
from google.colab import drive
import plotly.express as px

# REFRESH TOKEN para poder volver a realizar gets
import requests

url = "https://api.mercadolibre.com/oauth/token"

# Copiar el último refresh_token del output de esta celda y remplazarlo por el siguiente
payload = 'grant_type=refresh_token&client_id=6511763805540892&client_secret=wFdNuZNj1lgis0qYCCfFPKlVLXhiyzON&refresh_token=TG-68342bc8b623c60001af4a8f-232792475'
headers = {
  'accept': 'application/json',
  'content-type': 'application/x-www-form-urlencoded'
}

response = requests.request("POST", url, headers=headers, data=payload)

print(response.text)


# Utilizo el recurso de Predictor de categorías para obtener la categoría de productos buscados
# Lista de términos de búsqueda
busquedas = ['celular iphone', 'cafetera', 'televisor', 'notebook', 'calefactor', 'Colchonetas', 'campera montagne', 'monitor%20pc']
site = 'MLA'

# Token de acceso (si es necesario para autenticación)
headers = {
    'Authorization': f'Bearer {access_token}'  # solo si tu token es requerido
}

# Lista para almacenar resultados
resultados = []

# Iterar sobre los términos
for termino in busquedas:
    query = requests.utils.quote(termino)
    url = f"https://api.mercadolibre.com/sites/{site}/domain_discovery/search?q={query}"

    response = requests.get(url, headers=headers)
    data = response.json()
    category_id = data[0]["category_id"]
    category_name = data[0]["category_name"]
    resultados.append({
        "busqueda": termino,
        "category_id": category_id,
        "category_name": category_name
    })
    print(f"✅ {termino} → {category_id} ({category_name})")

# Convertir a DataFrame y guardar en CSV
df_resultados = pd.DataFrame(resultados)

category_ids = [item['category_id'] for item in resultados]

# Utilizo el recurso de Más vendidos por categoría para obtener el id de productos
# Lista de category_id (obtenidos previamente)

# Token de acceso (si es necesario para autenticación)
headers = {
    'Authorization': f'Bearer {access_token}'  # solo si tu token es requerido
}

# Almacenar los productos destacados por categoría
destacados = []

for categoria in category_ids:
    url = f"https://api.mercadolibre.com/highlights/{site}/category/{categoria}"
    response = requests.get(url, headers=headers)
    data = response.json()
    productos = data.get("content", [])
    for producto in productos:
        destacados.append({
            "category_id": categoria,
            "item_id": producto.get("id"),
            "position": producto.get("position")
        })

# Convertir a DataFrame y guardar
df_destacados = pd.DataFrame(destacados)

df_destacados
